+++ jest.config.js
@@ -2,6 +2,7 @@ module.exports = {
   preset: 'react-native',
   modulePathIgnorePatterns: [
     '<rootDir>/example/node_modules',
+    '<rootDir>/integration_test/node_modules',
     '<rootDir>/lib/',
   ],
 };